ANY_SRC en cours...
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@14220 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/schedulers/sched_opt/include/nm_so_raw_interface.h
@@ -47,7 +47,7 @@ nm_so_ri_swait_range(struct nm_so_interface *p_so_interface,
 
 extern int
 nm_so_ri_irecv(struct nm_so_interface *p_so_interface,
-	       uint16_t gate_id, uint8_t tag,
+	       long gate_id, uint8_t tag,
 	       void *data, uint32_t len,
 	       nm_so_request *p_request);
